refactor: resolver 클래스 분리
